requirements.txt
python-telegram-bot
APScheduler==3.0.0
psycopg2
sqlalchemy